requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
Django==4.2.16
django-appconf==1.0.6
django-geocoder==0.2.1
django-ipware==7.0.1
django-libsass==0.9
django-tinymce==4.0.0
django_compressor==4.5
djangorestframework==3.15.2
geocoder==1.15.1
gunicorn==22.0.0
idna==3.7
libsass==0.23.0
packaging==24.1
pillow==10.3.0
python-ipware==3.0.0
ratelim==0.1.6
rcssmin==1.1.2
requests==2.32.3
rjsmin==1.2.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
urllib3==2.2.2
user-agents==2.2.0
